fix: resolve circular dependency between dom.js and createList
