+++ src/utils/index.js
@@ -1,0 +1,4 @@
+export { dom } from './dom';
+export { sync } from './sync';
+export { reactive } from './reactive';
+export { generateUniqueId } from './nanoid';
